Day09_Visual_Question_Answering/requirements.txt
torch
transformers
pillow
requests
streamlit